Removed oot_level_writer, contents moved into appropriate files
